pintar estado del usuario en dropdown y unir nombre y apellido
