+++ src/views/Formulario/Formulario.js
@@ -1,6 +1,7 @@
 import Card from './components/Card'
 import InformacionPersonal from './components/InformacionPersonal'
 import NavPerfil from './../../components/perfil/NavPerfil'
+import { Header } from '../../components/Header'
 import {Footer} from "./../../components/Footer"
 
 
@@ -8,7 +9,7 @@ const Formulario = () => {
     return(
         <div className="">
             <div className="mb-5">
-                <NavPerfil />
+                <Header />
             </div>
             <Card />
             <InformacionPersonal />
